business_lookup_tool/business_lookup_tool/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# Data processing
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
folium==0.14.0

# Geospatial
geopy==2.3.0

# Caching and performance
Flask-Caching==2.0.2
redis==4.6.0

# Security
Flask-Limiter==3.5.0
Flask-Talisman==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Deployment
whitenoise==6.5.0
